refactor: rename telemetry module to logging
This matches the purpose of those modules better.

Closes #104
