+++ scheduler/src/lib.rs
@@ -1,3 +1,3 @@
 pub mod adapters;
 pub mod core;
-pub mod telemetry;
+pub mod logging;
